[metronome] enh: consolidate Metronome alert adapter
Moves the default/per-user spend-limit Metronome alert helpers
into a single adapter, updates callers and tests to import it
directly, and removes the old split modules.

This is a refactor-only slice: no product/API behavior changes intended.
